backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Celery & Redis
celery==5.3.4
redis==5.0.1

# Firebase
firebase-admin==6.4.0

# DigitalOcean Spaces (S3-compatible)
boto3==1.34.34

# Existing dependencies
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
yt-dlp==2023.12.30
crewai==0.1.0
groq==0.4.0

# Additional
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
